Move the `doc-server` module under the `@bayou` namespace.
