+++ local-modules/@bayou/doc-server/BaseDataManager.js
