+++ local-modules/@bayou/doc-server/CaretColor.js
