+++ local-modules/@bayou/doc-server/FileBootstrap.js
