+++ local-modules/@bayou/doc-server/SnapshotManager.js
