+++ local-modules/@bayou/doc-server/mocks/MockControl.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { DurableControl } from 'doc-server';
+import { DurableControl } from '@bayou/doc-server';
 import { MockSnapshot } from '@bayou/ot-common/mocks';
 
 /**
