+++ local-modules/@bayou/doc-server/tests/test_BaseDataManager.js
@@ -6,14 +6,14 @@ import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
 import { TheModule as appCommon_TheModule } from '@bayou/app-common';
-import { BaseDataManager, FileAccess, ValidationStatus } from 'doc-server';
+import { BaseDataManager, FileAccess, ValidationStatus } from '@bayou/doc-server';
 import { TransactionSpec } from '@bayou/file-store-ot';
 import { MockFile } from '@bayou/file-store/mocks';
 
 /** {FileAccess} Convenient instance of `FileAccess`. */
 const FILE_ACCESS = new FileAccess(appCommon_TheModule.modelCodec, new MockFile('blort'));
 
-describe('doc-server/BaseDataManager', () => {
+describe('@bayou/doc-server/BaseDataManager', () => {
   describe('.initSpec', () => {
     it('should call through to the impl', () => {
       const spec = new TransactionSpec();
