+++ local-modules/app-setup/DebugTools.js
@@ -8,7 +8,7 @@ import { inspect } from 'util';
 
 import { TheModule as appCommon_TheModule } from '@bayou/app-common';
 import { DocumentId } from '@bayou/doc-common';
-import { DocServer } from 'doc-server';
+import { DocServer } from '@bayou/doc-server';
 import { AuthorId } from '@bayou/ot-common';
 import { Logger } from '@bayou/see-all';
 import { RecentSink } from '@bayou/see-all-server';
