+++ local-modules/app-setup/RootAccess.js
@@ -4,7 +4,7 @@
 
 import { SplitKey } from '@bayou/api-common';
 import { Context } from '@bayou/api-server';
-import { DocServer } from 'doc-server';
+import { DocServer } from '@bayou/doc-server';
 import { Hooks } from '@bayou/hooks-server';
 import { Logger } from '@bayou/see-all';
 import { TString } from '@bayou/typecheck';
