+++ local-modules/app-setup/package.json
@@ -12,7 +12,7 @@
     "@bayou/deps-express": "local",
     "@bayou/deps-util-common": "local",
     "@bayou/doc-common": "local",
-    "doc-server": "local",
+    "@bayou/doc-server": "local",
     "@bayou/env-server": "local",
     "@bayou/hooks-server": "local",
     "@bayou/ot-common": "local",
